-Edited AudioManager -Added Default Game Settings Values -Added Clamp To Get Selected Color In CarMaterialManager -Edited Launcher, MenuManager -Changed Player Car Color Properties From Int To Byte For Performance -Selected Car Colors Now Get Saved

+++ Assets/Resources/Scripts/Settings/PlayerProperties.cs
@@ -183,7 +183,7 @@ namespace Andrich.UtilityScripts
 
     public static class CarMaterialExtensions
     {
-        public static void SetPrimaryMaterialIndex(this Player player, int index)
+        public static void SetPrimaryMaterialIndex(this Player player, byte index)
         {
             PhotonHashtable primaryIndex = new PhotonHashtable();  // using PUN's implementation of Hashtable
             primaryIndex[PlayerProperties.SelectedPrimaryMaterialProperty] = index;
@@ -191,10 +191,10 @@ namespace Andrich.UtilityScripts
             player.SetCustomProperties(primaryIndex);  // this locally sets the primary material index and will sync it in-game asap.
         }
 
-        public static void ChangePrimaryMaterialIndex(this Player player, int amount)
+        public static void ChangePrimaryMaterialIndex(this Player player, byte amount)
         {
-            int current = player.GetPrimaryMaterialIndex();
-            current = current + amount;
+            byte current = player.GetPrimaryMaterialIndex();
+            current = (byte)(current + amount);
 
             PhotonHashtable primaryIndex = new PhotonHashtable();  // using PUN's implementation of Hashtable
             primaryIndex[PlayerProperties.SelectedPrimaryMaterialProperty] = current;
@@ -202,18 +202,18 @@ namespace Andrich.UtilityScripts
             player.SetCustomProperties(primaryIndex);  // this locally sets the primary index and will sync it in-game asap.
         }
 
-        public static int GetPrimaryMaterialIndex(this Player player)
+        public static byte GetPrimaryMaterialIndex(this Player player)
         {
             object primaryIndex;
             if (player.CustomProperties.TryGetValue(PlayerProperties.SelectedPrimaryMaterialProperty, out primaryIndex))
             {
-                return (int)primaryIndex;
+                return (byte)primaryIndex;
             }
 
             return 0;
         }
 
-        public static void SetSecondaryMaterialIndex(this Player player, int index)
+        public static void SetSecondaryMaterialIndex(this Player player, byte index)
         {
             PhotonHashtable secondaryIndex = new PhotonHashtable();  // using PUN's implementation of Hashtable
             secondaryIndex[PlayerProperties.SelectedSecondaryMaterialProperty] = index;
@@ -221,10 +221,10 @@ namespace Andrich.UtilityScripts
             player.SetCustomProperties(secondaryIndex);  // this locally sets the secondary material index and will sync it in-game asap.
         }
 
-        public static void ChangeSecondaryMaterialIndex(this Player player, int amount)
+        public static void ChangeSecondaryMaterialIndex(this Player player, byte amount)
         {
-            int current = player.GetSecondaryMaterialIndex();
-            current = current + amount;
+            byte current = player.GetSecondaryMaterialIndex();
+            current = (byte)(current + amount);
 
             PhotonHashtable secondaryIndex = new PhotonHashtable();  // using PUN's implementation of Hashtable
             secondaryIndex[PlayerProperties.SelectedSecondaryMaterialProperty] = current;
@@ -233,12 +233,12 @@ namespace Andrich.UtilityScripts
 
         }
 
-        public static int GetSecondaryMaterialIndex(this Player player)
+        public static byte GetSecondaryMaterialIndex(this Player player)
         {
             object secondaryIndex;
             if (player.CustomProperties.TryGetValue(PlayerProperties.SelectedSecondaryMaterialProperty, out secondaryIndex))
             {
-                return (int)secondaryIndex;
+                return (byte)secondaryIndex;
             }
 
             return 0;
